fix(imports): Downloadliste blieb leer — Fliesskommawerte brachen die Antwort — v0.77.1
Drei Fehler, die zusammen dafuer sorgten, dass laufende Downloads unsichtbar
blieben und stattdessen ein roter Serverfehler stand:

**Fliesskommawerte.** yt-dlp liefert Groesse, Tempo und Restzeit als
Fliesskommazahlen ("239690509.7142857"), der Server reichte sie unveraendert
durch. Die App-Felder sind ganzzahlig — das Parsen der Antwort scheiterte
komplett, mitsamt der ganzen Liste. Der Server rundet jetzt (Bytes und
Sekunden haben keine Nachkommastellen), und die App nimmt zusaetzlich
Fliesskommazahlen an, damit ein einzelnes Komma nie wieder die gesamte Antwort
unbrauchbar macht.

**Verschluckter Fehler.** loadPending() hatte nur einen Erfolgszweig. Der
Parse-Fehler verschwand damit spurlos: leere Liste, kein Hinweis, kein Grund.
Jetzt landet er sichtbar in der Fehlerzeile.

**HTTP 500 im Archiv.** Der Aktualisieren-Knopf ruft POST /channels/:id/poll,
und fuer den manual-Kanal versuchte der Server einen RSS-Abruf auf
"manual:hikari" — keine Feed-URL, also 500. Das Archiv kennt keinen Abruf: Es
waechst nur durch eigene Importe. Der Endpunkt antwortet jetzt entsprechend,
statt zu scheitern.

Backend 654 Tests, Android 124 Tests, alle gruen.

+++ android/app/src/main/java/com/hikari/app/data/api/dto/ImportVideosDto.kt
@@ -134,10 +134,14 @@ data class PendingImportDto(
     @SerialName("isMovie") val isMovie: Boolean = false,
     /** "queued" | "downloading" | "failed" */
     val status: String,
-    @SerialName("downloadedBytes") val downloadedBytes: Long = 0,
-    @SerialName("totalBytes") val totalBytes: Long? = null,
-    @SerialName("speedBps") val speedBps: Long? = null,
-    @SerialName("etaSeconds") val etaSeconds: Int? = null,
+    // Bewusst Double statt Long/Int: yt-dlp liefert Groesse, Tempo und Restzeit
+    // als Fliesskommazahlen. Der Server rundet sie inzwischen, aber ein
+    // ganzzahliges Feld liesse die gesamte Antwort scheitern, sobald doch
+    // einmal ein Komma durchkommt — und mit ihr die ganze Downloadliste.
+    @SerialName("downloadedBytes") val downloadedBytes: Double = 0.0,
+    @SerialName("totalBytes") val totalBytes: Double? = null,
+    @SerialName("speedBps") val speedBps: Double? = null,
+    @SerialName("etaSeconds") val etaSeconds: Double? = null,
     @SerialName("fragmentIndex") val fragmentIndex: Int? = null,
     @SerialName("fragmentCount") val fragmentCount: Int? = null,
     /** 0…1, oder null solange sich der Anteil nicht bestimmen lässt. */
